Fix 281
* Delete unused import

* Delete unused import

* 增加新的背景色

* Delete unused import

* 改成 weui 样式

* 修改成 weui 样式

* Revert "增加新的背景色"

This reverts commit 17fc2b0dfdad500070dc85b0ed5b3b3293969a42.

* Revert "Revert "增加新的背景色""

This reverts commit 5208b0e6bd42fb3eb446728d448bea30cd2b7fd1.

* Revert "改成 weui 样式"

This reverts commit 5913565094dc3ac0fd004d5299986ae18f63bb89.

* Revert "Revert "改成 weui 样式""

This reverts commit a2b98f121ecbec0476c88f4c75ea2ae637f40f83.

* Revert "Delete unused import"

This reverts commit 77681633bb333e795b8ff21cfa02aec09650ce68.

* Delete unused import

* Revert "Revert "Revert "改成 weui 样式"""

This reverts commit 34d967b51d75b0a80d81fdef027bf02fdb1410e1.

* Revert "Revert "Revert "Revert "改成 weui 样式""""

This reverts commit b0f2c4222e1081c8305638779804b6f5c72f5d32.

* 提交订单时增加备注功能

* 更改路由 & Del Unuse imp

* 便签选择器 => 类选择器

* 增加‘备注’字段

* 更改ref

* 增加支付功能

* git 抽风了
