+++ app/actions/OrderActions.js
@@ -45,7 +45,7 @@ OrderActions.list.listen(function(type,state){
 OrderActions.get.listen(function(id){
   var data = {
     Id : id,
-    Fields : 'Id,UserId,BuyerName,BuyerTel,AppointedTime,PhotographerId,AlbumsId,CreationTime,State,Photographer.BusinessPhone,Photographer.NickName,Photographer.Avatar,User.NickName,User.Avatar,User.Id,Albums.Title,Albums.Cover,Albums.Price,Albums.Service,Albums.Description,PaymentTime,Price,CompleteTime,HasRefund,Refund.CompletionTime,Refund.Compensation'
+    Fields : 'Id,UserId,BuyerName,BuyerTel,AppointedTime,PhotographerId,AlbumsId,CreationTime,State,Photographer.BusinessPhone,Photographer.NickName,Photographer.Avatar,User.NickName,User.Avatar,User.Id,Albums.Title,Albums.Cover,Albums.Price,Albums.Service,Albums.Description,PaymentTime,BuyerMemo,Price,CompleteTime,HasRefund,Refund.CompletionTime,Refund.Compensation'
   };
   HttpFactory.post(API.ORDER.get,data,this.success,this.failed);
 });
