+++ app/scss/layout/_OrderDetailLayout.scss
@@ -23,7 +23,7 @@
       i { font-size: 32px; }
     }
 	}
-  /*p:last-child {
+  .last-layout {
     height: 60px;
     line-height: 18px;
 
@@ -39,5 +39,5 @@
         text-overflow: ellipsis;
       }
     }
-  }*/
+  }
 }
